[frontend] add tab fintel template
rename FintelTemplatesGrid to FintelTemplatesManager and remove the grid in that component
