+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/fintel_templates/FintelTemplateForm.tsx
@@ -107,6 +107,7 @@ const FintelTemplateForm = ({
             {!isEdition && (
               <FormButtonContainer>
                 <Button
+                  variant="secondary"
                   disabled={isSubmitting}
                   onClick={() => {
                     handleReset();
