+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/fintel_templates/useFintelTemplateAdd.tsx
@@ -1,6 +1,6 @@
 import { graphql } from 'react-relay';
 import { useState } from 'react';
-import { fintelTemplatesFragmentParams } from '@components/settings/sub_types/fintel_templates/FintelTemplatesGrid';
+import { fintelTemplatesFragmentParams } from '@components/settings/sub_types/fintel_templates/FintelTemplatesManager';
 import { useFintelTemplateAddMutation } from './__generated__/useFintelTemplateAddMutation.graphql';
 import useApiMutation from '../../../../../utils/hooks/useApiMutation';
 import { insertNodeFromEdge } from '../../../../../utils/store';
