+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/fintel_templates/useFintelTemplateDelete.tsx
@@ -1,6 +1,6 @@
 import { graphql, UseMutationConfig } from 'react-relay';
 import { useState } from 'react';
-import { fintelTemplatesFragmentParams } from '@components/settings/sub_types/fintel_templates/FintelTemplatesGrid';
+import { fintelTemplatesFragmentParams } from '@components/settings/sub_types/fintel_templates/FintelTemplatesManager';
 import useApiMutation from '../../../../../utils/hooks/useApiMutation';
 import { useFintelTemplateDeleteMutation } from './__generated__/useFintelTemplateDeleteMutation.graphql';
 import { deleteNodeFromEdge } from '../../../../../utils/store';
